requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
sqlalchemy>=1.4.0
alembic>=1.7.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-multipart>=0.0.5
aioredis>=2.0.0
redis>=4.0.0
prometheus-client>=0.12.0
prometheus-fastapi-instrumentator>=5.7.1
httpx>=0.19.0
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1
locust>=2.8.0
strawberry-graphql>=0.96.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
pika>=1.2.0
opentelemetry-api>=1.7.1
opentelemetry-sdk>=1.7.1
opentelemetry-instrumentation-fastapi>=0.26b1
jaeger-client>=4.8.0
email-validator>=1.1.3 